Move 'stac' serializers
