+++ django-rgd-imagery/rgd_imagery/serializers/__init__.py
@@ -1,6 +1,7 @@
 from .base import ImageMetaSerializer, ImageSerializer, ImageSetSerializer
 from .processed import ConvertedImageSerializer, RegionImageSerializer
 from .raster import RasterMetaSerializer, RasterSerializer
+from .stac import STACRasterSerializer
 
 __all__ = [
     'ConvertedImageSerializer',
@@ -10,4 +11,5 @@ __all__ = [
     'RasterMetaSerializer',
     'RasterSerializer',
     'RegionImageSerializer',
+    'STACRasterSerializer',
 ]
